refactor: keep history listener in Root
